+++ spp_drims_sl/tests/__init__.py
@@ -1,2 +1,3 @@
 # Part of OpenSPP. See LICENSE file for full copyright and licensing details.
 from . import test_spp_drims_sl
+from . import test_demo_user_safety