+++ spp_grm_demo/tests/__init__.py
@@ -2,3 +2,4 @@
 
 from . import test_grm_demo_wizard
 from . import test_story_tickets
+from . import test_demo_user_safety